Move passt network binding plugin to cmd/sidecars
