+++ BUILD.bazel
@@ -193,7 +193,7 @@ container_bundle(
         "$(container_prefix)/$(image_prefix)example-disk-mutation-hook-sidecar:$(container_tag)": "//cmd/sidecars/disk-mutation:example-disk-mutation-hook-sidecar-image",
         "$(container_prefix)/$(image_prefix)example-cloudinit-hook-sidecar:$(container_tag)": "//cmd/sidecars/cloudinit:example-cloudinit-hook-sidecar-image",
         "$(container_prefix)/$(image_prefix)network-slirp-binding:$(container_tag)": "//cmd/sidecars/network-slirp-binding:network-slirp-binding-image",
-        "$(container_prefix)/$(image_prefix)network-passt-binding:$(container_tag)": "//cmd/network-passt-binding:network-passt-binding-image",
+        "$(container_prefix)/$(image_prefix)network-passt-binding:$(container_tag)": "//cmd/sidecars/network-passt-binding:network-passt-binding-image",
         # container-disk images
         "$(container_prefix)/$(image_prefix)alpine-container-disk-demo:$(container_tag)": "//containerimages:alpine-container-disk-image",
         "$(container_prefix)/$(image_prefix)cirros-container-disk-demo:$(container_tag)": "//containerimages:cirros-container-disk-image",
@@ -343,7 +343,7 @@ container_push(
 container_push(
     name = "push-network-passt-binding",
     format = "Docker",
-    image = "//cmd/network-passt-binding:network-passt-binding-image",
+    image = "//cmd/sidecars/network-passt-binding:network-passt-binding-image",
     registry = "$(container_prefix)",
     repository = "$(image_prefix)network-passt-binding",
     tag = "$(container_tag)",
