+++ cmd/sidecars/network-passt-binding/BUILD.bazel
@@ -3,10 +3,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["main.go"],
-    importpath = "kubevirt.io/kubevirt/cmd/network-passt-binding",
+    importpath = "kubevirt.io/kubevirt/cmd/sidecars/network-passt-binding",
     visibility = ["//visibility:private"],
     deps = [
-        "//cmd/network-passt-binding/server:go_default_library",
+        "//cmd/sidecars/network-passt-binding/server:go_default_library",
         "//pkg/hooks:go_default_library",
         "//pkg/hooks/info:go_default_library",
         "//pkg/hooks/v1alpha2:go_default_library",
